Interfaces for tabs edited
