+++ client/src/components/tabbar/data/BookTabs.tsx
@@ -2,17 +2,8 @@
 import BackupTableIcon from '@mui/icons-material/BackupTable';
 import BookIcon from '@mui/icons-material/Book';
 import { Typography } from '@mui/material';
+import { TabInterface } from '../interfaces/TabInterfaces';
 
-
-export interface TabInterface {
-    icon:JSX.Element;
-    label: string;
-}
-
-export interface TabPanelInterface {
-    value: number;
-    items: JSX.Element[];
-}
 
 export const BookTabs:TabInterface[] = [
     {icon : (<BackupTableIcon />), label: "Table View" },
