+++ client/src/components/tabbar/interfaces/TabInterfaces.ts
@@ -1,0 +1,14 @@
+/**
+ * Tab Interface
+ */
+export interface TabInterface {
+    icon:JSX.Element;
+    label: string;
+}
+/**
+ * Tab Panel Interface
+ */
+export interface TabPanelInterface {
+    value: number;
+    items: JSX.Element[];
+}
